implemented old defaultUsersList method and renamed the current one
